+++ src/extension.js
@@ -1,7 +1,21 @@
 const vscode = require("vscode");
 const crypto = require("crypto");
-const { exec } = require("child_process");
+const { spawn } = require("child_process");
 
+/**
+ * @type {{ readonly CAPNP_TOOL_PATH: string, readonly IMPORT_PATH: string[], readonly STANDARD_IMPORT: string}}
+ */
+const CFG = Object.defineProperties({}, {
+	CAPNP_TOOL_PATH: {
+		get: () => vscode.workspace.getConfiguration("capnp").get("tool.path"),
+	},
+	IMPORT_PATH: {
+		get: () => vscode.workspace.getConfiguration("capnp").get("importPath"),
+	},
+	STANDARD_IMPORT: {
+		get: () => vscode.workspace.getConfiguration("capnp").get("standardImport"),
+	}
+});
 
 /**
  * Activate the extension.
@@ -21,14 +35,40 @@ function activate(context) {
 
 /**
  * 
+ * @param  {...string} args 
+ * @returns {Promise<{error: import("child_process").ExecException | null, code: number | null, stdout: string, stderr: string}>}
+ */
+async function capnpExec(...args) {
+	return new Promise((res, rej) => {
+		let stdout = "";
+		let stderr = "";
+
+		const CAPNP_BIN = CFG.CAPNP_TOOL_PATH || "capnp";
+
+		console.log(CAPNP_BIN, ...args);
+
+		const capnp = spawn(CAPNP_BIN, [...args], {
+			stdio: ["ignore", "pipe", "pipe"]
+		});
+		capnp.stdout.on('data', data => {
+			stdout += data;
+		});
+		
+		capnp.stderr.on('data', data => {
+			stderr += data;
+		});
+
+		capnp.on("error", error => res({ code: null, error, stdout, stderr }));
+		capnp.on('close', code => res({ code: code, error: null, stdout, stderr }));
+	});
+}
+
+/**
+ * 
  * @returns {Promise<boolean>} Whether or not capnp could be retrieved.
  */
-async function tryGetCapnp() {
-	return new Promise(res => {
-		exec("capnp --version", (error, stdout) => {
-			res(!error);
-		})
-	})
+function isCapnpReachable() {
+	return capnpExec("--version").then(x => x.code == 0 && x.error == null);
 }
 
 /**
@@ -36,14 +76,16 @@ async function tryGetCapnp() {
  * @param {vscode.ExtensionContext} context 
  */
 async function initCapnp(context) {
-	const available = await tryGetCapnp();
+	const available = await isCapnpReachable();
 	if (!available) {
-		vscode.window.showErrorMessage(
-			"Unable to execute the capnp command line tool.\n" +
-			"Please ensure it is installed and in the PATH."
+		console.warn("Capnp tool not found.");
+		vscode.window.showWarningMessage(
+			"Unable to execute the capnp tool.\n" +
+			"Please ensure it is installed and in PATH.\n" +
+			"\n" +
+			"Code diagnostics will be unavailable."
 		);
 	}
-	console.log("Capnp tool is reachable");
 
 	const diagnostics = vscode.languages.createDiagnosticCollection("capnp");
 	const changeWatcher = vscode.workspace.onDidSaveTextDocument((e) => {
@@ -96,17 +138,20 @@ function parseCompileErrors(str) {
 function capnpDocumentChanged(uri, diagnostics) {
 	console.log("File changed - running diagnostics");
 
-	// TODO: The fspath should be escaped.
-	exec(`capnp compile ${uri.fsPath} -o-`, (error, _, stderr) => {
-		const compileErrors = error ? stderr : "";
-		const newDiags = parseCompileErrors(compileErrors).map(err =>
-			new vscode.Diagnostic(
-				new vscode.Range(err.rowStart, err.colStart, err.rowEnd, err.colEnd),
-				err.message
-			)
-		);
-		diagnostics.set(uri, newDiags);
-	});
+	const standardImport = CFG.STANDARD_IMPORT ? [] : ["--no-standard-import"];
+	const imports = CFG.IMPORT_PATH.map(x => "-I" + x);
+
+	capnpExec("compile", ...standardImport, ...imports, uri.fsPath, "-o-")
+		.then(({code, stderr}) => {
+			const compileErrors = code != 0 ? stderr : "";
+			const newDiags = parseCompileErrors(compileErrors).map(err =>
+				new vscode.Diagnostic(
+					new vscode.Range(err.rowStart, err.colStart, err.rowEnd, err.colEnd),
+					err.message
+				)
+			);
+			diagnostics.set(uri, newDiags);
+		});
 }
 
 function generateUid(...args) {
@@ -117,7 +162,7 @@ function generateUid(...args) {
 	const upper = uidArray[0].toString(16);
 	const lower = uidArray[1].toString(16).padStart(8, "0");
 	const uid = `@0x${upper}${lower}`;
-	insertIntoEditor(uid)
+	insertIntoEditor(uid);
 }
 
 /**
@@ -127,8 +172,8 @@ function generateUid(...args) {
 async function insertIntoEditor(str) {
 	await vscode.window.activeTextEditor.insertSnippet({
 		value: str
-	})
+	});
 }
 
 
-module.exports = { activate }
+module.exports = { activate };